[DEBUG] correct line generations

+++ esvg/Base.cpp
@@ -433,7 +433,7 @@ void esvg::Base::setId(const std::string& _newId) {
 
 void esvg::Base::drawShapePoints(std::vector<std::vector<vec2>>& _out,
                                  int32_t _recurtionMax,
-                                 int32_t _threshold,
+                                 float _threshold,
                                  mat2& _basicTrans,
                                  int32_t _level) {
