+++ esvg/Base.h
@@ -83,7 +83,7 @@ namespace esvg {
 			 */
 			virtual void drawShapePoints(std::vector<std::vector<vec2>>& _out,
 			                             int32_t _recurtionMax,
-			                             int32_t _threshold,
+			                             float _threshold,
 			                             mat2& _basicTrans,
 			                             int32_t _level=1);
