+++ esvg/Polygon.cpp
@@ -134,7 +134,7 @@ void esvg::Polygon::draw(esvg::Renderer& _myRenderer, mat2& _basicTrans, int32_t
 
 void esvg::Polygon::drawShapePoints(std::vector<std::vector<vec2>>& _out,
                                     int32_t _recurtionMax,
-                                    int32_t _threshold,
+                                    float _threshold,
                                     mat2& _basicTrans,
                                     int32_t _level) {
 	ESVG_VERBOSE(spacingDist(_level) << "DRAW Shape esvg::Polygon");
