+++ esvg/Polyline.cpp
@@ -131,7 +131,7 @@ void esvg::Polyline::draw(esvg::Renderer& _myRenderer, mat2& _basicTrans, int32_
 
 void esvg::Polyline::drawShapePoints(std::vector<std::vector<vec2>>& _out,
                                      int32_t _recurtionMax,
-                                     int32_t _threshold,
+                                     float _threshold,
                                      mat2& _basicTrans,
                                      int32_t _level) {
 	ESVG_VERBOSE(spacingDist(_level) << "DRAW Shape esvg::Polyline");
