+++ esvg/esvg.cpp
@@ -1,8 +1,6 @@
 /** @file
  * @author Edouard DUPIN
- * 
  * @copyright 2011, Edouard DUPIN, all right reserved
- * 
  * @license APACHE v2.0 (see license file)
  */
 
@@ -390,7 +388,7 @@ std::vector<std::vector<vec2>> esvg::Document::getLines(vec2 _size) {
 
 void esvg::Document::drawShapePoints(std::vector<std::vector<vec2>>& _out,
                                      int32_t _recurtionMax,
-                                     int32_t _threshold,
+                                     float _threshold,
                                      mat2& _basicTrans,
                                      int32_t _level) {
 	ESVG_VERBOSE(spacingDist(_level) << "DRAW shape esvg::Document");
